% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeAUC}
\alias{computeAUC}
\title{Calculate AUC from a fitted object}
\usage{
computeAUC(fit, dmin, dmax, islogd = TRUE)
}
\arguments{
\item{fit}{usually a drFit object. However, any object with a predict method would work.}

\item{dmin}{minimum dose. The integral range is [dmin, dmax].}

\item{dmax}{maximum dose. The integral range is [dmin, dmax].}

\item{islogd}{whether the supplied dose dmin/dmax is in log10 scale. The user should be responsible for the consistency between the actual value
of dmin, dmax and islogd. If log10 transformed dmin/dmax is supplied with islogd=TRUE, the AUC is calculated based on dose-response curve with x-axis being log10(dose); On the
other hand, if dmin, dmax is original scale and islogd=FALSE, the AUC is calculated based on a dose-response curve with x-axis being dose}
}
\value{
a vector of AUC, AUC0 and AUCs. AUC is the area under dose response curve; AUC0 is the area under the line response=1; AUCS is AUC/AUC0
}
\description{
AUC is calculated through the integrate() function based on dose-response curve.
}

