% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer.R
\name{computeICvariance}
\alias{computeICvariance}
\title{compute variance of IC value at each inhibition percentage (original dose scale)}
\usage{
computeICvariance(drFit, percent = 0.5)
}
\arguments{
\item{drFit}{A drFit object as returned by drFit() function.}

\item{percent}{the inhibition ratio to be searched against. A vector between 0 and 1. Corresponding
response is 1-percent}
}
\value{
a vector of estimated variance of the doses achieving each response level
}
\description{
Notice that this does not allow log.d option for log10 dose; it also does not allow interpretation/extrapolation. When
a response is not achieved, the variance estimate would be NA indicating not achieved
}

