% Generated by roxygen2 (4.0.2): do not edit by hand
\name{detect_ray_design}
\alias{detect_ray_design}
\title{detect if a experiment is fixed-ray design and prepare input for CI computation}
\usage{
detect_ray_design(d1, d2, e, tol = 0.001, d2.d1.force = NA)
}
\arguments{
\item{d1}{d1}

\item{d2}{d2}

\item{e}{e}

\item{tol}{tolerance in declaring fixed ratio.}

\item{d2.d1.force}{the ratio d2/d1 to be forced in fitting the linear model; this is only effective if the data is not a fixed ratio design
usually, the all-possible combination design; specifying d2.d1.force will use the specified ratio (from available ratios) to fit the model and generate dose-response curve.}
}
\value{
a list
}
\description{
detect if a experiment is fixed-ray design and prepare input for CI computation
}

