% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drexplorerGUI_1}
\alias{drexplorerGUI_1}
\title{Launch Graphical User Interface (GUI) for dose response curve fitting}
\usage{
drexplorerGUI_1()
}
\description{
This function will launch GUI for curve fitting. The GUI works across different platforms, but the appearance would be slightly different.
}

