% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drexplorerGUI_2}
\alias{drexplorerGUI_2}
\title{Launch Graphical User Interface (GUI) for drug interaction (IAI) analysis}
\usage{
drexplorerGUI_2()
}
\description{
This function will launch GUI for computing Interaction Index (IAI) in fixed ratio design.
The GUI works across different platforms, but the appearance would be slightly different.
}

