% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drOutlier}
\alias{drOutlier}
\title{Identifying outliers for dose-response data}
\usage{
drOutlier(drMat, alpha = 0.01)
}
\arguments{
\item{drMat}{dose-response matrix. the first column being dosage and second column being response.
controls are included by specifying dose=0}

\item{alpha}{a scalar for significance level. Only 0.01, 0.05 and 1 are allowed.
alpha=1 is included for comparability issue. In this case,  no outliers will be identified.}
}
\value{
indicator a vector with length nrow(drMat) specifying if a data point is outlier.
 Control points always have status FALSE
}
\description{
A wrapper for the NewmanTest() function. It repeatedly call NewmanTest() for each dose level.
}
\examples{
data(ryegrass, package='drc') # use the ryegrass data from drc package
drOutlier(drMat=ryegrass[, c(2, 1)], alpha=0.05)
}
\author{
Kevin R Coombes (\email{kcoombes@mdanderson.org}), Pan Tong (\email{nickytong@gmail.com})
}
\seealso{
\code{\link{NewmanTest}, \link{drModels}, \link{drFit}, \link{drFit-class}}
}

