% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_hillNIC60.R
\name{fHill}
\alias{fHill}
\title{the Hill equation function}
\usage{
fHill(Dose, Einf, E0, logEC50, HS)
}
\arguments{
\item{Dose}{the concentration of drug (original scale)}

\item{Einf}{the bottom asymptote of response (e.g. at drug concentration of Inf)}

\item{E0}{the top asymptote of response (e.g. at drug concentration of 0)}

\item{logEC50}{natural logarithm of EC50}

\item{HS}{Hill slope}
}
\value{
the response at dose=Dose (between 0 and 1)
}
\description{
the Hill equation function
}

