% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSA_Bliss_Raydesign.R
\name{fitHSA_ray}
\alias{fitHSA_ray}
\title{fit HSA and Bliss model given a fixed ratio design with input (d1, d2, e)}
\usage{
fitHSA_ray(d1, d2, e, tol = 0.05, strict = TRUE, name1 = "Drug A",
  name2 = "Drug B", dmin = NULL, dmax = NULL, islogd = TRUE,
  bootstrapB = 0, nrep = 3, seed = 100)
}
\arguments{
\item{d1}{drug 1 dose in original scale}

\item{e}{response at the dose pair}

\item{strict}{whether to stop if not fixed ratio; this is just for visualization; the calculations of synergy is wrong since d12=d1+d2 is meaningless for variable ratio!
At the end, this is a waste of time: it would not make sense and impossible to do this!}

\item{dmin}{start position to calcualte AUC at projected dose, in original scale, default is minimum projected dose}

\item{dmax}{end position to calcualte AUC at projected dose, in original scale, default is maximum projected dose}

\item{islogd}{whether to compute AUC at log10 dose; if TRUE, dmin and dmax must also be log10 scale}

\item{d1}{drug 2 dose in original scale}
}
\description{
fit HSA and Bliss model given a fixed ratio design with input (d1, d2, e)
}

