% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSA_Bliss_Raydesign.R
\name{fitHSA_ray_boot}
\alias{fitHSA_ray_boot}
\title{bootstrap of fitHSA_ray to obtain variation of deltaAUC}
\usage{
fitHSA_ray_boot(d1, d2, e, tol = 0.05, strict = TRUE, name1 = "Drug A",
  name2 = "Drug B", dmin = NULL, dmax = NULL, islogd = TRUE,
  bootstrapB = 10, nrep = 3, seed = 100)
}
\arguments{
\item{bootstrapB}{number of boostrap iterations}

\item{nrep}{how many replicates are sampled with replacement for each dose combination. nrep=3 will make it much narrower CI band}
}
\description{
bootstrap of fitHSA_ray to obtain variation of deltaAUC
}

