% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer.R
\docType{methods}
\name{lines,drFit-method}
\alias{lines,drFit-method}
\title{lines method for drFit object}
\usage{
\S4method{lines}{drFit}(x, col = 5, lwd = 2, lty = 1,
  show_points = FALSE, pcol = "black", pch = 16, type = c("line"), ...)
}
\arguments{
\item{x}{a drFit object}

\item{col}{line color}

\item{lwd}{line width}

\item{show_points}{whether to add points for dose-response paires (dose not 0)}

\item{pcol}{color for points; only relevant if show_points is TRUE}

\item{pch}{pch for points; only relevant if show_points is TRUE}

\item{type}{the plot type. The user can specify more flexibly by specifying a vector of entities to plot including:
line: dose response curve
points: observed data points}

\item{...}{additional parametrs passed to generic lines() function}
}
\description{
lines method for drFit object
}

