% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_hillNIC60.R
\name{lines.hillFit}
\alias{lines.hillFit}
\title{lines method for hillFit class}
\usage{
\method{lines}{hillFit}(fit, col = 1, lwd = 2, show_points = FALSE,
  pcol = "black", pch = 16, ...)
}
\arguments{
\item{col}{line color}

\item{lwd}{line width}

\item{show_points}{whether to add points for dose-response paires (dose not 0)}

\item{pcol}{color for points; only relevant if show_points is TRUE}

\item{pch}{pch for points; only relevant if show_points is TRUE}

\item{...}{additional parametrs passed to generic lines() function}

\item{x}{a hillFit object as returned by hillFit}
}
\description{
lines method for hillFit class
}

