% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_hillNIC60.R
\name{lines.nci60Fit}
\alias{lines.nci60Fit}
\title{lines method for nci60Fit class}
\usage{
\method{plot}{nci60Fit}(fit, pcol = "black", lcol = "black", lwd = 2,
  lty = 1, pch = 16, type = c("line"))
}
\arguments{
\item{x}{a nci60Fit object as returned by nci60Fit}

\item{h}{horizontal line to add indicating e.g. GI (h=0.5), TGI (h=0), LD50 (h=-0.5)}

\item{...}{additional parameters, not implemented}
}
\description{
lines method for nci60Fit class
}

