% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadData}
\alias{loadData}
\title{read data; if extension is csv, treat it as csv; otherwise, treat it as tab delimilated file through read.delim}
\usage{
loadData(filepath)
}
\description{
read data; if extension is csv, treat it as csv; otherwise, treat it as tab delimilated file through read.delim
}

