% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nci60Fit}
\alias{nci60Fit}
\title{implementation of NCI60 method (GI50, TGI, LC50) for dose response data}
\usage{
nci60Fit(d, y, interpolation = FALSE)
}
\arguments{
\item{d}{dose original dose}

\item{y}{response percentage of growth inhibition as defined by NCI60 method}
}
\description{
Notice that here the response is assumed to be calculated as in: http://dtp.nci.nih.gov/branches/btb/ivclsp.html. The response has range
-1 to 1.
}
\references{
Shoemaker, R. H. The NCI60 Human Tumour Cell line Anticancer Drug Screen. Nature Reviews, 6: 813-823, 2006.
}

