\name{NewmanTables}
\alias{onePercentTab}
\alias{fivePercentTab}
\docType{data}
\title{Table III and IV published by Newman}
\description{
This contain two 20-by-12 matrices corresponding to table III and table IV published by Newman, D. In
particular, onePercentTab gives the quantiles at significance level of 0.01 and fivePercentTab gives the quantiles at significance level of 0.01.
The degree of freedom f and sample size n are used as row and column names for each table. 
}

\format{
  Each of the two objects (onePercentTab, fivePercentTab)
  is a matrix with 20 rows (degree of freedom f) and 12 columns (sample size n).
}

\references{
Newman, D. (1939). The distribution of range in samples from a normal population, 
expressed in terms of an independent estimate of standard deviation. Biometrika, 31(1/2), 20-30.
}
\seealso{
\code{\link{NewmanTest}}
}
\keyword{datasets}

