\name{nl22B2}
\alias{nl22B2}
\title{The drug combination data between SCH66336 and 4-HPR. This is an all-possible combination design.}
\format{A data frame with 30 rows and 3 columns}
\source{
  \url{https://biostatistics.mdanderson.org/SoftwareDownload/}
}
\description{
  This dataset is published in Lee et al, 2007, Journal of
  Biopharmaceutical Statistics, Table 1
}
\details{
  \itemize{ \item schd SCH66336 dose \item hpr 4-HPR dose
  \item resp response }
}
\references{
  Lee, J. Jack, et al. Interaction index and different
  methods for determining drug interaction in combination
  therapy. Journal of biopharmaceutical statistics 17.3
  (2007): 461-480.
}

