% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer.R
\docType{methods}
\name{plot,drFit-method}
\alias{plot,drFit-method}
\title{plot method for drFit object}
\usage{
\S4method{plot}{drFit}(x, pchs = c(16, 17, 15), cols = c(1, 2, 3),
  col = 4, lwd = 2, lty = 1, addLegend = FALSE, xlab = "Dose",
  ylab = "Relative viability", ylim = NA, xlim = NA, main,
  type = c("line", "points", "control", "legend", "sem"), style = "",
  bty = "n", h = c(0.5), cex.main = 1, cex.axis = 1, cex.lab = 1,
  axes = TRUE, return = FALSE)
}
\arguments{
\item{x}{a drFit object}

\item{pchs}{pchs (a vector) specify symbols to show different data points. In particular,
pchs[1] is used to show regular points. pchs[2] is used to show outliers at 0.05 significance level and pchs[3] is used 
to show outliers at 0.01 significance level.}

\item{cols}{Similarly, cols[1] is used to color regular points, cols[2] to color outliers at 0.05 significance level and
cols[3] to color outliers at 0.01 significance level.}

\item{col}{color used to specify the appearance of fitted curve}

\item{lwd}{line width used to specify the appearance of fitted curve}

\item{addLegend}{currently deprecated. Please specify through type argument. whether to add legend indicating outlier status}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{ylim}{y limit for display}

\item{xlim}{x limit for display}

\item{main}{main title}

\item{type}{the plot type. The user can specify more flexibly by specifying a vector of entities to plot including:
line: dose response curve
points: observed data points
control: control data points
legend: legend for outlier status}

\item{style}{deprecated. Please specify through type argument. when style='full', observed dose-response pairs are plotted including controls, fitted curve is superimposed and legend for outliers indicated;
when style='simple', only fitted curve is plotted (without dose-response points and of course, not outlier status).}

\item{bty}{bty passed to legend}

\item{h}{horizontal line to add indicating e.g. IC50 (h=0.5)}

\item{cex.main}{cex for main title}

\item{cex.axis}{cex for axis annotation}

\item{cex.lab}{cex for axis label}

\item{axes}{whether to add axes}

\item{return}{whether return data for the plotted lines; this is useful if we want to reconstruct the lines elsewhere}
}
\description{
plot method for drFit object
}

