% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_hillNIC60.R
\name{plot.hillFit}
\alias{plot.hillFit}
\title{plot method for hillFit class}
\usage{
\method{plot}{hillFit}(fit, xlab = "Log10(Dose)",
  ylab = "Relative viability", main = "Fit of Hill equation", xlim = NULL,
  ylim = NULL, cex.main = 1, cex.axis = 1, pcol = "black",
  lcol = "black", lwd = 2, ...)
}
\arguments{
\item{...}{additional parameters, not implemented}

\item{x}{a hillFit object as returned by hillFit}
}
\description{
plot method for hillFit class
}

