% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSA_Bliss_Raydesign.R
\name{plotHSA_ray}
\alias{plotHSA_ray}
\title{plot the ray design from fit object returned by fitHSA_ray}
\usage{
plotHSA_ray(fit, cols = c(1, 2, 3), xlim = NULL, ylim = NULL,
  xlab = NULL, ylab = "Relative viability", main = NULL, tag = NULL,
  type = c("line", "HSA", "Bliss", "se"), legend = TRUE, plotSetup = TRUE,
  digit = 3)
}
\arguments{
\item{cols}{color corresponding to mixture, drugA, drugB}

\item{xlim}{x-axis limit}

\item{ylim}{y-axis limit}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{legend}{whether to show legend}

\item{plotSetup}{whether initiate a new figure; set to False when just to add curve as lines}
}
\description{
plot the ray design from fit object returned by fitHSA_ray
}

