% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.hillFit}
\alias{predict.hillFit}
\title{predict method for hillFit class}
\usage{
\method{predict}{hillFit}(fit, newData = NULL)
}
\arguments{
\item{newData}{dose in original scale to be used for prediction of response}

\item{x}{a hillFit object as returned by hillFit}
}
\value{
predicted response at specified dose
}
\description{
predict method for hillFit class
}

