% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer.R
\name{prepDRdat}
\alias{prepDRdat}
\title{prepare dose-reponse data}
\usage{
prepDRdat(drMat, alpha = 0.01, fitCtr = FALSE, standardize = TRUE)
}
\arguments{
\item{drMat}{dose-response matrix as for drFit. the first column being dosage and second column being response. controls are included by specifying dose=0}

\item{alpha}{a scalar for significance level. This specifies the significance level to identify outliers which will be excluded from model fitting. To include
all data, set alpha=1.}

\item{fitCtr}{A logic vector specifying whether to include the control points into the model fitting.}

\item{standardize}{whether to standardize (scale) the data based on control points. This should be disabled when no control data is supplied}
}
\value{
a list
}
\description{
this function scaled the response by mean reponse in control when necessary
}
\details{
the standardization means response (e.g. count) is to be scaled by mean control response so that
the standardized response is relative viability, usually between 0 and 1.
this function first detects outlier data points (using supplied data, either scaled or not scaled). It
then scale the data and split the data into two data frames: one for dose not 0 and one for dose at 0 
Ideally, outlier detection is better at original count (or signal intensity) due to asymptotic assumption
But for data already scaled, detection outlier at this level is what we can do
}

