% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepDRdat2expInfo}
\alias{prepDRdat2expInfo}
\title{from prepared data to experiment setup information}
\usage{
prepDRdat2expInfo(resPrepDR)
}
\arguments{
\item{resPrepDR}{result returned by prepDRdat}
}
\value{
a data frame
}
\description{
this includes: N dose (non-control), N technical rep (number of obs. at each dose, may be a string),
experimental variation including medianSD, minSD, maxSD, meanSD computed from SD across different tech replicates
}

