% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer.R
\name{recommendedModels}
\alias{recommendedModels}
\title{frequently used models}
\usage{
recommendedModels()
}
\description{
this function returns frequently used dose-response models for anti-cancer drug screening.
}

