% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ryegrass}
\alias{ryegrass}
\title{ryegrass dataset from drc package}
\description{
This dataset is from the ryegrass data in drc package. It is a 2 column data frame where the conc column is the concentration of ferulic acid in mM
and rootl column is the root length of ryegrass.
}
\references{
Inderjit and J. C. Streibig, and M. Olofsdotter (2002) Joint action of phenolic acid mixtures and its
significance in allelopathy research, Physiologia Plantarum, 114, 422–428, 2002
}

