% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer.R
\name{toget}
\alias{toget}
\title{general getter function for attributes in an object}
\usage{
toget(obj, what = "avail")
}
\arguments{
\item{obj}{object with attributes}

\item{what}{specify what attribute to get}
}
\value{
attributes extracted
}
\description{
general getter function for attributes in an object
}

