% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorerAdded.R
\name{datSplit}
\alias{datSplit}
\title{Split a drug-response data}
\usage{
datSplit(dat, colNames, by, minReplicate = 3, verbose = TRUE)
}
\arguments{
\item{dat}{input data with multiple cell lines for one drug}

\item{colNames}{colNames to name selected columns}

\item{by}{string for the colName (just one column allowed) used for split}

\item{minReplicate}{if a subset does not have nrow exceeding this, throw an error}

\item{verbose}{whether to produce warning message}
}
\value{
a list of the split data frames
}
\description{
This split the data into a list so that each element is a list. This is useful if the data contains multiple cell lines or multiple drugs for the same cell line
Currently it only allows splitting by one variable (one column). If data contains both multiple cell lines and multiple drugs, a customized function needs to be written for data formatting.
}

