% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_1_source_v2.R
\name{drexplorerGUI_1}
\alias{drexplorerGUI_1}
\title{Launch Graphical User Interface (GUI) for dose response curve fitting}
\usage{
drexplorerGUI_1()
}
\description{
This function will launch GUI for curve fitting. The GUI works across different platforms, but the appearance would be slightly different.
}

