% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_2_source.R
\name{drexplorerGUI_2}
\alias{drexplorerGUI_2}
\title{Launch Graphical User Interface (GUI) for drug interaction (IAI) analysis}
\usage{
drexplorerGUI_2()
}
\description{
This function will launch GUI for computing Interaction Index (IAI) in fixed ratio design. 
The GUI works across different platforms, but the appearance would be slightly different.
}

