% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drFit}
\alias{drFit}
\title{Fit a dose-response model}
\usage{
drFit(drMat, modelName = "sigEmax", alpha = 0.01, fitCtr = FALSE,
  standardize = TRUE)
}
\arguments{
\item{drMat}{dose-response matrix. the first column being dosage and second column being response. controls are included by specifying dose=0}

\item{modelName}{a dose-response model. For available models to be specified, see drModels().}

\item{alpha}{a scalar for significance level. This specifies the significance level to identify outliers which will be excluded from model fitting. To include
all data, set alpha=1.}

\item{fitCtr}{A logic vector specifying whether to include the control points into the model fitting.}

\item{standardize}{whether to standardize (scale) the data based on control points. This should be disabled when no control data is supplied}
}
\value{
the function returns a drFit S4 object.
}
\description{
This function can fit various dose-response models by specifying a model name and package source (either drc or DoseFinding).
}
\details{
This is a wrapper to fit dose response models from drc and DoseFinding packages.
When fit the model, the response is internally scaled with mean response in control. Therefore, the fitted response is a ratio.
In visualization, i.e. dose-response curve plotting, the dose is log10 transformed so that we can see the sigmoid curve.
However, in computing IC50 and doing prediction, the dose is in original scale since the model is trained in the original scale of dose. Correspondingly, predicted
response is also scaled response.
When the user requires a log10 transformed dose, it is done after estimating dose in original scale through the computeIC() function.
}
\examples{
data(ryegrass, package='drc') # use the ryegrass data from drc package
# fit a sigmaEmax model without outlier removal. the controls are excluded from model fitting
fit.LL.3 <- drFit(drMat=ryegrass[, c(2, 1)], modelName = "LL.3", alpha=0.01, fitCtr=FALSE)
fit.LL.3u <- drFit(drMat=ryegrass[, c(2, 1)], modelName = "LL.3u", alpha=0.01, fitCtr=FALSE)
fit.sigEmax <- drFit(drMat=ryegrass[, c(2, 1)], modelName = "sigEmax", alpha=0.01, fitCtr=FALSE)
}
\author{
Kevin R Coombes (\email{kcoombes@mdanderson.org}), Pan Tong (\email{nickytong@gmail.com})
}
\seealso{
\code{\link{NewmanTest}, \link{drOutlier}, \link{drModels}, \link{drFit-class}}
}

