% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionIndex.R
\name{fitIAI}
\alias{fitIAI}
\title{Estimate the interaction index as well as its confidence interval using delta method for fixed ratio design}
\usage{
fitIAI(d1, d2, e, E = seq(0.05, 0.95, 0.005), name1 = "Drug A",
  name2 = "Drug B", alpha = 0.05, d2.d1.force = NA, tol = 0.01)
}
\arguments{
\item{d1}{dose for drug 1}

\item{d2}{dose for drug 2}

\item{e}{corresponding response in the range [0, 1]}

\item{E}{a vector of responses (between 0 and 1) where IAI and confidence interval are to be computed from.}

\item{name1}{name of drug 1}

\item{name2}{name of drug 2}

\item{alpha}{significance level of confidence interval}

\item{d2.d1.force}{a ratio passed to detect_ray_design() function so as to specify a fixed ratio for grid design}
}
\value{
a data frame with columns IAI, IAI.low, IAI.up, E, dx1 (corresponding dose of drug 1), dx2 (corresponding dose of drug 1),
}
\description{
this code is extracted from the source code distributed at https://biostatistics.mdanderson.org/SoftwareDownload/
}
\details{
Two papers have been published by Lee et al, one in 2007 (Lee2007) and on in 2009 (Lee2009). The Lee2007 paper 
described five methods to assess interaction: (1) Lowewe additivity model using interaction index (IAI) (2) Model of Greco et al 1990.
This approach uses \deqn{\alpha}{alpha} as the metric and it can be related to IAI (3) Model of Machado and Robinson which uses a metric denoted
as \deqn{\eta}{eta} (4) Model of Plummer and Short which can also be linked to IAI through the parameter \deqn{\beta_4}{beta_4} (5) Model of
Carter et al that can be linked to IAI through the parameter \deqn{\beta_{12}}{beta_12}. For more details of these models, please refer to Lee2007.

The Lee2009 paper provided generalization of IAI to multiple drugs using Lowewe additivity model and assumption of Chou and Talalay's median effect
equation. The Chou and Talalay's median effect equation can be expressed as: \deqn{log(E/(1-E))=m(log d - log Dm)}{log(E/(1-E))=m(log d - log Dm)} 
where E is the effect at dose d for a
compound whose median effective dose.

Some notes about experiment design. Usually the data is either fixed ratio design (ray design) or grid design which means all-possible combination of drug concentrations
between two drugs are available. The Lee2007 paper provided an example of grid design. However, specific fixed ratio is used to fit the median effect model which is the basis to estimate IAI. The Lee2009 paper
considered with fixed ratio design.
}
\references{
Lee, J. J., & Kong, M. (2009). Confidence intervals 
  of interaction index for assessing multiple drug interaction. Statistics in biopharmaceutical research, 1(1), 4-17.

Lee, J. Jack, et al (2007). Interaction index and different 
   methods for determining drug interaction in combination therapy. Journal of biopharmaceutical statistics 17.3 461-480.
}

