% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hillFit}
\alias{hillFit}
\title{fit Hill equation on dose response data}
\usage{
hillFit(d, y)
}
\arguments{
\item{d}{dose original dose}

\item{y}{response relative viability}
}
\description{
Notice that control data (dose=0) is usually not fitted (supplied) in Hill equation.
}

