% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{lines,drFit-method}
\alias{lines,drFit-method}
\title{lines method for drFit object}
\usage{
\S4method{lines}{drFit}(x, col = 5, lwd = 2, show_points = FALSE,
  pch = 16)
}
\arguments{
\item{x}{a drFit object}

\item{col}{line color (also applies to data points if show_points is TRUE}

\item{lwd}{line width}

\item{show_points}{whether to add points for dose-response paires (dose not 0)}

\item{pch}{pch for points}
}
\description{
lines method for drFit object
}

