% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_median_effect}
\alias{plot_median_effect}
\title{generate plot for the fit from median-effect model}
\usage{
plot_median_effect(medianEffect, type = c("medianEffect", "doseResponseCurve",
  "contour"), contour.level = (1:9)/10)
}
\arguments{
\item{medianEffect}{the fitted result by fit_median_efect()}

\item{type}{type of plot, options are from c('medianEffect', 'doseResponseCurve', 'contour')}

\item{contour.level}{levels (between 0 and 1 representing different response values) to draw contour plot}
}
\description{
generate plot for the fit from median-effect model
}

