% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionIndex.R
\name{plot_median_effect}
\alias{plot_median_effect}
\title{generate plot for the fit from median-effect model}
\usage{
plot_median_effect(medianEffect, type = c("medianEffect", "doseResponseCurve",
  "contour"), contour.level = (1:9)/10, logd = FALSE, cex.legend = 1,
  align = TRUE, ylim = c(0, 1), legend.position = "topright")
}
\arguments{
\item{medianEffect}{the fitted result by fit_median_efect()}

\item{type}{type of plot, options are from c('medianEffect', 'doseResponseCurve', 'contour')}

\item{contour.level}{levels (between 0 and 1 representing different response values) to draw contour plot}

\item{logd}{whether to plot logd in Dose-response curve; only effective if type=='doseResponseCurve'}

\item{cex.legend}{cex for the legend text'}

\item{align}{whether to align dose range. Imagine that 2 drugs are mixed in fixed ratio, the doses in one drug will have low doses that the other drug and mixture do not have 
which some biologists deams as a waste of graphic region. An easy fix is to truncate xlim. this is only only effective if type=='doseResponseCurve'}

\item{legend.position}{legend position passed to legend() function}
}
\description{
generate plot for the fit from median-effect model
}

