% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionIndex.R
\name{plotIAI}
\alias{plotIAI}
\title{Visualize interaction index}
\usage{
plotIAI(fit, type = c("IAI", "medianEffect", "doseResponseCurve", "contour"),
  contour.level = (1:9)/10, ylim = NULL, mode = "both", logd = FALSE,
  align = TRUE, cex.legend = 1, legend.position = "topright")
}
\arguments{
\item{fit}{the fitted result from fitIAI()}

\item{type}{type of plot from c('IAI', 'medianEffect', 'doseResponseCurve', 'contour')}

\item{contour.level}{contour level. only effective if type='contour'}

\item{ylim}{y axis limit}

\item{mode}{specify if to plot against response, dose or both. only effective if type=='IAI'. can be either 'response', 'dose', or 'both'}

\item{logd}{logd passed to plot_median_effect}

\item{align}{align passed to plot_median_effect}

\item{cex.legend}{cex.legend passed to plot_median_effect}

\item{legend.position}{legend.position passed to plot_median_effect}
}
\description{
Visualize interaction index
}

