% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer.R
\docType{methods}
\name{predict,drFit-method}
\alias{predict,drFit-method}
\title{predict method for drFit object}
\usage{
\S4method{predict}{drFit}(object, newData)
}
\arguments{
\item{object}{a drFit object}

\item{newData}{a vector of dose levels to predict at. Default is to predict response at observed dose levels.}
}
\value{
a vector of predicted values
}
\description{
predict method for drFit object
}

