% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.nci60Fit}
\alias{predict.nci60Fit}
\title{predict method for nci60Fit class}
\usage{
\method{predict}{nci60Fit}(fit, newData = NULL)
}
\arguments{
\item{newData}{dose in original scale to be used for prediction of response}

\item{x}{a nci60Fit object as returned by nci60Fit}
}
\value{
predicted response at specified dose
}
\description{
predict method for nci60Fit class
}

