% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionIndex.R
\name{UMSCC22B}
\alias{UMSCC22B}
\title{The drug combination data of squamous cell carcinoma cells (UMSCC22B) surviving after 72 hours of treatment by single and
combination dose levels of SCH66336 and 4-HPR. This is a fixed-ratio design.}
\format{A data frame with 13 rows and 3 columns}
\source{
\url{https://biostatistics.mdanderson.org/SoftwareDownload/}
}
\description{
This dataset is published in Lee et al, 2009, Statistics in biopharmaceutical research, Table 2
}
\details{
\itemize{
  \item dose1 SCH66336 dose
  \item dose2 4-HPR dose
  \item e response
}
}
\references{
Lee, J. J., & Kong, M. (2009). Confidence intervals 
  of interaction index for assessing multiple drug interaction. Statistics in biopharmaceutical research, 1(1), 4-17.
}

